pyproject.toml
[tool.poetry]
name = "wordle_bot"
version = "0.1.0"
description = "LangGraph to play Wordle-like game"
authors = ["Lakshya Agarwal <lakshyajannu@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.8.0"
langchain-core = "^0.2.38"
langchain-openai = "^0.1.23"
langchain-community = "^0.2.16"
langgraph = "^0.2.18"
python-dotenv = "^1.0.1"
langsmith = "^0.1.115"
typer = "^0.12.5"

[tool.poetry.scripts]
wordle_bot = "wordle_bot.guess:app"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
